+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 * the length of the tags has been increased to 40 characters;
 * added `VersionUpdatesCommand`, that performs some updates to the database or
     files needed for versioning;
+* improved tag validation for posts. Now errors on multiple tags are shown at the
+    same time and the tag name that generated the error is shown;
+* `PostValidator::validTagsLength()` and `PostValidator::validTagsChars()` have
+    been replaced with the `PostValidator::validTags()` method. `TagValidatorTrait`
+    has been removed and its methods has been moved into `TagValidor` class;
 * fixed a bug when loading plugins, when the `Asset` plugin does not exist.
 
 ### 2.26.0